backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
ddgs==9.5.4
Jinja2==3.1.4
tenacity==9.0.0
openai==1.51.2
email-validator==2.2.0
python-dotenv==1.0.0
langgraph